Cancel button "works"
Cancelling the folder transfer is now possible using the progress dialog cancel button. The way this is done is really bad. the thread gets terminated immediately resulting in the file being transferred to be stuck in use by the program. This results in the inability to delete that file. This also means there will be two folders with the same name in different parent folders. (Hide smaller folder?)